+++ packages/elements-utils/src/index.ts
@@ -1,0 +1,2 @@
+export { generateToC, injectHttpOperationsAndModels } from './toc';
+export * from './toc/types';